% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equal_var.R
\name{bptable}
\alias{bptable}
\title{Prints out Breusch-Pagan Test and commentary if specified}
\usage{
bptable(data, x, y, commentary = NULL)
}
\arguments{
\item{data}{the dataset name}

\item{x}{the x values}

\item{y}{the y values}

\item{commentary}{a string of interpretation for Breusch-Pagan Test}
}
\value{
a summary table and interpretation
}
\description{
Prints out Breusch-Pagan Test and commentary if specified
}
