% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independence.R
\name{dwtable}
\alias{dwtable}
\title{Printing out Durbin-Watson test and commentary if specified}
\usage{
dwtable(data, x, y, commentary = NULL)
}
\arguments{

\item{data}{a data frame containing the variables}

\item{x}{the independent variable name}

\item{y}{the dependent variable name}

\item{commentary}{a logical value indicating to print interpretations}

}
\value{
a Durbin-Watson test and interpretation
}
\description{
Printing out Durbin-Watson test and commentary if specified
}
