% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equal_var.R
\name{equalvar_commentary}
\alias{equalvar_commentary}
\title{Prints out Breusch-Pagan Test interpretation}
\usage{
equalvar_commentary(bp_values)
}
\arguments{
\item{bp_values}{the model's Breusch-Pagan values}
}
\value{
an interpretation of results
}
\description{
Prints out Breusch-Pagan Test interpretation
}
