% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independence.R
\name{independence_commentary}
\alias{independence_commentary}
\title{Printing out Durbin-Watson test and commentary if specified}
\usage{
independence_commentary(dw_value)
}
\arguments{
\item{dw_value}{Durbin-Watson values from model}
}
\value{
a Durbin-Watson test interpretation
}
\description{
Printing out Durbin-Watson test and commentary if specified
}
