% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearity.R
\name{linearity_commentary}
\alias{linearity_commentary}
\title{Writes adjusted r-squared and p-value interpretations}
\usage{
linearity_commentary(model)
}
\arguments{
\item{model}{the linear model}
}
\value{
interpretations
}
\description{
Writes adjusted r-squared and p-value interpretations
}
