% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearity.R
\name{linearplot}
\alias{linearplot}
\title{Makes a scatter plot to determine if the data follows a linear pattern}
\usage{
linearplot(data, x, y, title)
}
\arguments{
\item{data}{the dataset name}

\item{x}{the x values}

\item{y}{the y values}

\item{title}{the title}
}
\value{
a scatterplot with a regression line
}
\description{
Makes a scatter plot to determine if the data follows a linear pattern
}
