% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normality.R
\name{normality_commentary}
\alias{normality_commentary}
\title{Printing out Shapiro-Wilk test commentary}
\usage{
normality_commentary(sw_values)
}
\arguments{
\item{sw_values}{Shapiro-Wilk test values from model}
}
\value{
a string of interpretations
}
\description{
Printing out Shapiro-Wilk test commentary
}
