% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normality.R
\name{qqgraph}
\alias{qqgraph}
\title{Makes a QQ Plot to Determine Normality}
\usage{
qqgraph(data, x, y)
}
\arguments{
\item{data}{the dataset name}

\item{x}{the x values}

\item{y}{the y values}
}
\value{
a QQ plot
}
\description{
Makes a QQ Plot to Determine Normality
}
