% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equal_var.R
\name{residplot}
\alias{residplot}
\title{Makes a residual plot to determine if the residuals demonstrate homoscedasticity}
\usage{
residplot(data, x, y)
}
\arguments{
\item{data}{the dataset name}

\item{x}{the x values}

\item{y}{the y values}
}
\value{
a residual plot
}
\description{
Makes a residual plot to determine if the residuals demonstrate homoscedasticity
}
