% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearity.R
\name{stats_table}
\alias{stats_table}
\title{Collects summary statistics for model and prints interpretations}
\usage{
stats_table(model, commentary = NULL)
}
\arguments{
\item{model}{the linear model}

\item{commentary}{a string of interpretations}
}
\value{
a summary table and interpretation
}
\description{
Collects summary statistics for model and prints interpretations
}
