% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normality.R
\name{swtable}
\alias{swtable}
\title{Printing out Shapiro-Wilk test and commentary if specified}
\usage{
swtable(data, x, commentary = NULL)
}
\arguments{
\item{data}{the dataset name}

\item{x}{the x values}

\item{commentary}{a string of interpretations}
}
\value{
a summary table with Shapiro-Wilk test
}
\description{
Printing out Shapiro-Wilk test and commentary if specified
}
