% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independence.R
\name{acf_pacf}
\alias{acf_pacf}
\title{Makes an ACF and PACF graph for both the X and Y variables}
\usage{
acf_pacf(data, x, y)
}
\arguments{
\item{data}{the dataset name}

\item{x}{the x values}

\item{y}{the y values}
}
\value{
An ACF and PACF graph for both the X and Y variables
}
\description{
Makes an ACF and PACF graph for both the X and Y variables
}
