% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independence.R
\name{dwtable}
\alias{dwtable}
\title{Printing out Durbin-Watson test and commentary if specified}
\usage{
dwtable(data, x, y, commentary = NULL)
}
\arguments{
\item{data}{the dataset name}

\item{x}{the x values}

\item{y}{the y values}

\item{commentary}{a string of interpretations}
}
\value{
a Durbin-Watson test and interpretation
}
\description{
Printing out Durbin-Watson test and commentary if specified
}
