% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independence.R
\name{independence_commentary}
\alias{independence_commentary}
\title{Helper function for commentary}
\usage{
independence_commentary(dw_value)
}
\arguments{
\item{dw_value}{Durbin-Watson values from model}
}
\value{
a Durbin-Watson test interpretation
}
\description{
Helper function for commentary
}
