% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normality.R
\name{qqgraph}
\alias{qqgraph}
\title{Makes a QQ Plot and does the Shapiro-Wilk's Test to Determine Normality}
\usage{
qqgraph(data, x, y)
}
\arguments{
\item{data}{the dataset name}

\item{x}{the x values}

\item{y}{the y values}

\item{title}{the title}
}
\value{
a scatterplot with a regression line
}
\description{
Makes a QQ Plot and does the Shapiro-Wilk's Test to Determine Normality
}
